Refactor endpoint functionality to be in aio.py
Remove EndpointBase, and change endpoint classes
to just store the appropriate endpoints.

Move all API functions to aio module.
